requirements.txt
PyQt5-5.15.2
pyperclip-1.8.2